move formatFlagsToArgs
