+++ packages/cli-common/src/index.ts
@@ -1,7 +1,7 @@
 export * from './lib/plugins/model'
 export { HookName, HookFunc, HooksListeners, Hooks } from './lib/hooks'
 export { PluginContext, PluginInitContext } from './lib/plugins/context'
-export { composeFlags, envIdFlags, tunnelServerFlags, urlFlags } from './lib/flags'
+export { composeFlags, envIdFlags, tunnelServerFlags, urlFlags } from './lib/common-flags'
+export { formatFlagsToArgs } from './lib/flags'
 export { initHook } from './hooks/init/load-plugins'
 export { default as BaseCommand } from './commands/base-command'
-export * as utils from './utils'
