+++ packages/cli/src/commands/proxy/urls.ts
@@ -1,5 +1,5 @@
 import { ux, Args } from '@oclif/core'
-import { tunnelServerFlags, urlFlags, utils } from '@preevy/cli-common'
+import { tunnelServerFlags, urlFlags, formatFlagsToArgs } from '@preevy/cli-common'
 import { commands } from '@preevy/core'
 import { pick } from 'lodash'
 import PreevyUrlsCmd from '../urls'
@@ -36,7 +36,7 @@ export default class Urls extends ProfileCommand<typeof Urls> {
     if (!envId) {
       throw new Error('Proxy not running, use preevy proxy connect <compose-project>')
     }
-    const commandArgs = [`--id=${envId}`, ...utils.formatFlagsToArgs(flags, PreevyUrlsCmd.flags), ...Object.values(pick(this.args, Object.keys(PreevyUrlsCmd.args))).map(x => `${x}`)]
+    const commandArgs = [`--id=${envId}`, ...formatFlagsToArgs(flags, PreevyUrlsCmd.flags), ...Object.values(pick(this.args, Object.keys(PreevyUrlsCmd.args))).map(x => `${x}`)]
     await this.config.runCommand('urls', commandArgs)
 
     return undefined
